Organize code into package
