+++ pyproject.toml
@@ -27,8 +27,8 @@ dependencies = [
 ]
 
 [project.scripts]
-face-detect = "face_detection:main"
-face-app = "app:main"
+face-detect = "reconhecimento_facial.face_detection:main"
+face-app = "reconhecimento_facial.app:main"
 
 [tool.pytest.ini_options]
 addopts = "-vv"
