+++ reconhecimento_facial/emotion_detection.py
