+++ reconhecimento_facial/llm_service.py
