+++ reconhecimento_facial/obstruction_detection.py
