+++ tests/test_caption.py
@@ -4,7 +4,7 @@ from unittest import mock
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
-import llm_service
+import reconhecimento_facial.llm_service as llm_service
 
 
 def test_generate_caption(monkeypatch):
